RDCC-958: moved ProfessionalApiGeneratorConstants to constants package
