+++ src/test/java/uk/gov/hmcts/reform/professionalapi/controller/constants/ProfessionalApiGeneratorConstantsTest.java
@@ -1,10 +1,10 @@
-package uk.gov.hmcts.reform.professionalapi.generator;
+package uk.gov.hmcts.reform.professionalapi.controller.constants;
 
 import static org.junit.Assert.assertFalse;
 
 import java.lang.reflect.Constructor;
 import org.junit.Test;
-import uk.gov.hmcts.reform.professionalapi.generator.ProfessionalApiGeneratorConstants;
+import uk.gov.hmcts.reform.professionalapi.controller.constants.ProfessionalApiGeneratorConstants;
 
 public class ProfessionalApiGeneratorConstantsTest {
